backend/requirements.txt
Flask==3.1.1
Flask-Cors==4.0.0
boto3==1.34.137
python-dotenv==1.0.1
authlib==1.3.0
python-jose==3.3.0
